asr-model-service/requirements.txt
fastapi
uvicorn
transformers==4.41.1
torch
python-multipart
soundfile
pydub==0.25.1
python-magic
torchaudio